Add image upload for training, trainer and place
