+++ .gitignore
@@ -15,3 +15,5 @@ projects/open-training/public/bundles/
 projects/open-training/public/build/*/glyphicons-*
 
 .env
+
+public/uploads/images
