+++ projects/open-training/config/packages/twig.yaml
@@ -7,6 +7,7 @@ twig:
 
     globals:
         training_max_references_to_show: '%training_max_references_to_show%'
+        image_uploads: '%image_uploads%'
 
     # see https://symfony.com/blog/new-in-symfony-2-7-default-date-and-number-format-configuration
     date:
