+++ projects/open-training/packages/Training/src/Entity/Training.php
@@ -2,13 +2,11 @@
 
 namespace OpenTraining\Training\Entity;
 
-use DateTime;
 use DateTimeInterface;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\Mapping as ORM;
 use Gedmo\Mapping\Annotation as Gedmo;
-use Gedmo\Timestampable\Traits\Timestampable;
-use Symfony\Component\HttpFoundation\File\File;
+use OpenTraining\Entity\UploadableImageTrait;
 use Vich\UploaderBundle\Mapping\Annotation as Vich;
 
 /**
@@ -17,7 +15,7 @@ use Vich\UploaderBundle\Mapping\Annotation as Vich;
  */
 class Training
 {
-    use Timestampable;
+    use UploadableImageTrait;
 
     /**
      * @ORM\Id()
@@ -39,12 +37,6 @@ class Training
      * @var string
      */
     private $name;
-
-    /**
-     * @ORM\Column(type="string", length=255)
-     * @var string
-     */
-    private $image;
 
     /**
      * @ORM\Column(type="text")
@@ -75,12 +67,6 @@ class Training
      * @var int
      */
     private $price;
-
-    /**
-     * @Vich\UploadableField(mapping="training_images", fileNameProperty="image")
-     * @var File
-     */
-    private $imageFile;
 
     /**
      * @ORM\ManyToOne(targetEntity="OpenTraining\Training\Entity\Place")
@@ -277,33 +263,5 @@ class Training
         }
 
         return $this->getNearestTerm()->getSlug();
-    }
-
-    public function setImageFile(?File $file = null): void
-    {
-        $this->imageFile = $file;
-
-        // VERY IMPORTANT:
-        // It is required that at least one field changes if you are using Doctrine,
-        // otherwise the event listeners won't be called and the file is lost
-        if ($file) {
-            // if 'updatedAt' is not defined in your entity, use another property
-            $this->updatedAt = new DateTime('now');
-        }
-    }
-
-    public function getImageFile(): ?File
-    {
-        return $this->imageFile;
-    }
-
-    public function setImage(?string $image): void
-    {
-        $this->image = $image;
-    }
-
-    public function getImage(): ?string
-    {
-        return $this->image;
     }
 }
