+++ projects/open-training/templates/default/homepage.html.twig
@@ -15,8 +15,7 @@
     Welcome to lecture
 
     <ul>
-        <li><a href="{{ url('provision') }}">Provision compute</a></li>
-        <li><a href="{{ url('trainings') }}">Trainings</a></li>
+        <li><a href="{{ url('trainings') }}">Vzdělávej se</a></li>
         <li><a href="/admin/">Admin</a></li>
         <li><a href="/adminer?sqlite=&username=&db=%2Fvar%2Fwww%2Fopen-lectures%2Fvar%2Fdatabase.sqlite">Database</a></li>
     </ul>
